% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{AIC.movement_model}
\alias{AIC.movement_model}
\title{Akaike's An Information Criterion for the movement model oject}
\usage{
\method{AIC}{movement_model}(object, ...)
}
\arguments{
\item{object}{a \code{movement_model} object}

\item{\dots}{further arguments to be passed to or from other methods. Won't be used here.}
}
\description{
Calculate the AIC for the \code{movement_model} object provided.
}

