% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.data.frame.movement_matrix}
\alias{as.data.frame.movement_matrix}
\title{Convert a movement_matrix object into a data.frame}
\usage{
\method{as.data.frame}{movement_matrix}(movement_matrix)
}
\arguments{
\item{movement_matrix}{A \code{movement_matrix} object}
}
\value{
A standardised text string.
}
\description{
Convert a \code{movement_matrix} object into a \code{data.frame} with
columns \code{origin}(character), \code{destination}(character) and \code{movement}(numeric).
}
\note{
It will print a warning ... TODO
}

