% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.locationdataframe.data.frame}
\alias{as.locationdataframe.data.frame}
\title{Convert a merged data.frame into a single location data.frame}
\usage{
\method{as.locationdataframe}{data.frame}(input, ...)
}
\arguments{
\item{input}{A data.frame of the format
#   origin destination movement pop_origin  pop_destination
# 1      a           b       10        100               88
# 2      a           c        8        100              100
# 3      a           d       10        100              113
# 4      a           e       11        100              107
# 5      a           f        8        100               67
#   lat_origin long_origin  lat_destination long_destination
#   0.07826932  0.13612404       0.12114115       0.58984725
#   0.07826932  0.13612404       0.07126503       0.19544754
#   0.07826932  0.13612404       0.97817937       0.22771625
#   0.07826932  0.13612404       0.87233335       0.06695538
#   0.07826932  0.13612404       0.23157835       0.19573021}
}
\value{
A data.frame containing location data of the format
#   location pop        lat        lon
# 1        a 100 0.07826932 0.13612404
# 2        b  88 0.12114115 0.58984725
# 3        c 100 0.07126503 0.19544754
# 4        d 113 0.97817937 0.22771625
# 5        e 107 0.87233335 0.06695538
}
\description{
Takes a data.frame containing location and population data and converts it
into a single data.frame containing location data only.
}

