% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.locationdataframe}
\alias{as.locationdataframe}
\alias{as.locationdataframe.SpatialPolygonsDataFrame}
\alias{as.locationdataframe.data.frame}
\title{Conversion to locationdataframe}
\usage{
as.locationdataframe(input, ...)

\method{as.locationdataframe}{data.frame}(input, ...)

\method{as.locationdataframe}{SpatialPolygonsDataFrame}(input, populationraster,
  ...)
}
\arguments{
\item{input}{object to convert to a \code{locationdataframe} object.
Either a data.frame with columns \code{origin} (character), \code{destination} (character), \code{movement} (numeric),
\code{pop_origin} (numeric), \code{pop_destination} (numeric), \code{lat_origin} (numeric), \code{long_origin} (numeric),
\code{lat_destination} (numeric) and \code{long_destination} (numeric) or a \code{SpatialPolygonsDataFrame} object}

\item{populationraster}{a \code{RasterLayer} object with each the value of
each cell giving the associated population density.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A data.frame containing location data of the format
#   location pop        lat        lon
# 1        a 100 0.07826932 0.13612404
# 2        b  88 0.12114115 0.58984725
# 3        c 100 0.07126503 0.19544754
# 4        d 113 0.97817937 0.22771625
# 5        e 107 0.87233335 0.06695538
}
\description{
Convert objects to \code{locationdataframe} objects
}

