% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.movementmatrix}
\alias{as.movementmatrix}
\title{Conversion to locationdataframe}
\usage{
as.movementmatrix(dataframe)
}
\arguments{
\item{input}{object to convert to a \code{locationdataframe} object.
Either a data.frame with columns \code{origin} (character), \code{destination} (character), \code{movement} (numeric),
\code{pop_origin} (numeric), \code{pop_destination} (numeric), \code{lat_origin} (numeric), \code{long_origin} (numeric),
\code{lat_destination} (numeric) and \code{long_destination} (numeric) or a \code{SpatialPolygonsDataFrame} object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A data.frame containing location data with columns \code{location} (character), \code{pop} (numeric),
\code{lat} (numeric) and \code{lon} (numeric).
}
\description{
Convert objects to \code{locationdataframe} objects
}

