% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{createpopulationfromcsv}
\alias{createpopulationfromcsv}
\title{Create a data frame of populations at particular coordinates}
\usage{
createpopulationfromcsv(filename)
}
\arguments{
\item{filename}{File path of the csv file to process}
}
\value{
A dataframe containing csv data.
}
\description{
Reads a correctly formatted csv file and creates a dataframe
}

