% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geocodes}
\alias{geocodes}
\title{Geocodes}
\format{A data frame with 11385 rows and 3 variables:
\describe{
  \item{geoid}{int Geolocation id}
  \item{lon}{num longitude }
  \item{lat}{num latitude}
}}
\usage{
geocodes
}
\description{
Geocodes for places in popular movement data. Projection RT90 
  ESPG:2400.
}
\author{
Johan Junkka \email{johan.junkka@umu.se}
}
