% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{get.network}
\alias{get.network}
\title{Extract the necessary components for movement modelling form a population
density raster.}
\usage{
get.network(raster, min = 1, matrix = TRUE)
}
\arguments{
\item{raster}{A \code{RasterLayer} object of population density.}

\item{min}{The minimum population size for inclusion in the network. All
cells with populations greater than or equal to \code{min} will be included
and other excluded.}

\item{matrix}{Whether the distance matrix should be returned as a
\code{matrix} object (if \code{TRUE}) or as a \code{dist} object (if
\code{FALSE}).}
}
\value{
A list with three components:
 \item{population }{A vector giving the populations at the cells of
interest}
 \item{distance_matrix }{A distance matrix (either of class \code{matrix} or
\code{dist}) diving the pairwise euclidean distance between the cells of
interest in the units of \code{raster}}
 \item{coordinate }{A two-column matrix giving the coordinates of the cells
of interest in the units of \code{raster}}
}
\description{
Given raster map of population density, extract a distance matrix and vector
of population sizes for all cells with population density above a minimum
threshold. These can be used as network representation of the landscape for
use in movement models.
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# get the network for pixels with at least 50,000 inhabitants
net <- get.network(kenya10, min = 50000)
# visualise the distance matrix
sp::plot(raster::raster(net$distance_matrix))
# plot the raster layer
sp::plot(kenya10)
# rescale the population of those pixels for plotting
size <- 0.1 + 2 * net$population / max(net$population)
# plot the pixels selected, with point size proportional to population size
points(net$coordinates, pch = 16,
     cex = size,
     col = rgb(0, 0, 1, 0.6))
}
\seealso{
\code{\link{raster}}, \code{\link{dist}}
}

