% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{gravity}
\alias{gravity}
\title{Gravity model}
\usage{
gravity(theta = 0.01, alpha = 0.06, beta = 0.03, gamma = 0.01)
}
\arguments{
\item{theta}{Model parameter with default value and the limits theta = [0, Inf].}

\item{alpha}{Model parameter with default value and the limits alpha = [-Inf, Inf].}

\item{beta}{Model parameter with default value and the limits alpha = [-Inf, Inf].}

\item{gamma}{Model parameter with default value and the limits gamma = [-Inf, Inf].}
}
\value{
A flux model object with the \code{gravity flux} function and a set of starting parameters.
}
\description{
The gravity law assumes that the number of people moving between locations is
proportional to some power of the origin and destination population, and decays
by distance between them following:
\deqn{T_{ij} = \frac{m_i^\alpha \times n_j^\beta }{f(r_{ij})}}{T_ij = m_i^\alpha * n_j^\beta / f(r_ij)}
where, \eqn{m_i} represents the population at origin, \eqn{n_j} the population at the destination
and \eqn{r_{ij}}{r_ij} the distance between them. \eqn{\alpha} and \eqn{\beta} are tuning parameters
fitted to each subpopulation size, and \eqn{f(r_{ij})}{f(r_ij)} is a distance-dependent functional
form.
}
\note{
Limits \eqn{0} and \eqn{Inf} will be changed internally to the numerically safe approximations
\eqn{0 -> sqrt(.Machine$double.eps)} and \eqn{Inf -> sqrt(.Machine$double.xmax)}, respectively.
}
\references{
Zipf, G.K. (1946). The P1 P2 / D hypothesis: on the intercity movement of persons. \emph{Am. Sociol. Rev.},
11, 677-686.
Balcan, D., Colizza, V., Gonc, B. & Hu, H. (2009). Multiscale mobility networks and the spatial.
\emph{Proc. Natl. Acad. Sci. U. S. A.}, 106, 21484-9.
}
\seealso{
\code{\link{movement}}, \code{\link{original.radiation}}, \code{\link{radiation.with.selection}},
\code{\link{uniform.selection}}, \code{\link{intervening.opportunities}}, \code{\link{gravity.with.distance}}
}

