% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{gravity.with.distance}
\alias{gravity.with.distance}
\title{Gravity with distance model}
\usage{
gravity.with.distance(theta1 = 0.01, alpha1 = 0.06, beta1 = 0.03,
  gamma1 = 0.01, delta = 0.5, theta2 = 0.01, alpha2 = 0.06,
  beta2 = 0.03, gamma2 = 0.01)
}
\arguments{
\item{delta}{Model parameter with default value and the limits delta = [0, 1].}

\item{theta1/2}{Model parameter with default value and the limits theta = [0, Inf].}

\item{alpha1/2}{Model parameter with default value and the limits alpha = [-Inf, Inf].}

\item{beta1/2}{Model parameter with default value and the limits beta = [-Inf, Inf].}

\item{gamma1/2}{Model parameter with default value and the limits gamma = [-Inf, Inf].}
}
\value{
A flux model object with the \code{\link{gravitywithdistance.flux}} function and a set of starting
parameters.
}
\description{
In order to obtain more accurate results, following Viboud et al. 2006 we implement a nine-parameter
form of the gravity law, in which short and long trips are fitted separately. Similarly to the gravity
model we fit each parameter (equation 1) using a Poisson regression:
\deqn{T_{ij} = \theta \frac{ N_i^{\alpha} N_j^{\beta} }{d_{ij}^{\gamma}} }{%
T_ij = \theta * N_i^\alpha * N_j^{\beta} / d_ij^{\gamma} }
where \eqn{\theta} is a proportionality constant and the exponents \eqn{\alpha} and \eqn{\beta} respectively,
tune the dependence of dispersal on donor and recipient population sizes (\eqn{N}), and the distance between
the two communities \eqn{d_{ij}^{\gamma}}{d_ij^\gamma}. By taking the logarithm of on both sides this becomes:
\deqn{\ln(T_{ij}) = \ln(\theta) + \alpha \ln(N_i) + \beta \ln{N_j} - \gamma \ln(d_{ij}) }{%
 ln(T_ij) = ln(\theta) + \alpha ln(N_i) + \beta ln{N_j} - \gamma ln(d_ij)
 }
Viboud et al. show that below 119km, the population exponents are relatively high and larger for the
destination population. Therefore we allow the flexibility to adjust based on a distance cutoff for the model.
}
\note{
Limits \eqn{0} and \eqn{Inf} will be changed internally to the numerically safe approximations
\eqn{0 -> sqrt(.Machine$double.eps)} and \eqn{Inf -> sqrt(.Machine$double.xmax)}, respectively.
}
\references{
Viboud, C. et al. (2006). Synchrony, waves, and spatial hierarchies in the spread of influenza. \emph{Science},
312, 447-51
}
\seealso{
\code{\link{movement}}, \code{\link{gravitywithdistance.flux}}, \code{\link{original.radiation}},
\code{\link{radiation.with.selection}}, \code{\link{uniform.selection}}, \code{\link{intervening.opportunities}},
\code{\link{gravity}}
}

