% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{gravitywithdistance.flux}
\alias{gravitywithdistance.flux}
\title{Use the Simini et al. 2012 modified gravitation model to predict
movement between two sites.}
\usage{
gravitywithdistance.flux(i, j, distance, population, theta = c(1, 0.6, 0.3, 3,
  1, 1, 0.6, 0.3, 3), symmetric = FALSE, minpop = 1, maxrange = Inf)
}
\arguments{
\item{i}{Index for \code{population} and \code{distance} giving the first
site}

\item{j}{Index for \code{population} and \code{distance} giving the second
site}

\item{distance}{A distance matrix giving the euclidean distance between
pairs of sites}

\item{population}{A vector giving the population at all sites}

\item{theta}{A vector of nine parameters in the order: scalar1,
exponent1 on donor pop, exponent1 on recipient pop, exponent1 on distance,
theshhold, scalar2, exponent2 on donor pop, exponent2 on recipient pop,
exponent2 on distance. The first four parameters are used as the parameters
of \code{gravity.flux} if the distance is less than threshold (5th
parameter), otherwise the last four parameters are used for the gravity
flux.}

\item{symmetric}{Whether to return a single value giving the total predicted
movements from i to j and j to i (if \code{TRUE}) or vector of length 2
giving movements from i to j (first element) and from j to i (second element)}

\item{minpop}{The minimum population size to consider (by default 1, consider
all sites)}

\item{maxrange}{The maximum distance between sites to consider (by default
\code{Inf}, consider all sites)}
}
\value{
A vector (of length either 1 or 2) giving the predicted number of
people moving between the two sites.
}
\description{
Given indices \code{i} and \code{j}, a vector of population sizes
\code{population}, a (dense) distance matrix \code{distance} giving the
euclidean distances between all pairs of sites, and a set of parameters
\code{theta}, to predict movements between sites \code{i} and \code{j}.
The flux can be calculated either for both directions (by setting
 \code{symmetric = FALSE}, returning movements for each direction) or for
 the summed movement between the two (\code{symmetric = TRUE}).
The model can be sped up somewhat by setting \code{minpop} and
\code{maxrange}. If either of the two sites has a population lower than
\code{minpop} (minimum population size), or if the distance between the two
sites is greater than \code{maxrange} (the maximum range) it is assumed that
no travel occurs between these points.
Note that this function only works for individual sites, use
\code{\link{movement.predict}} to calculate movements for multiple
populations. The modification from Simini et al. introduces a nine parameter
form where a different set of parameters are used if the distance is greater
than \code{delta}.
}
\examples{
# generate random coordinates and populations
n <- 30
coords <- matrix(runif(n * 2), ncol = 2)
pop <- round(runif(n) * 1000)
# calculate the distance between pairs of sites
d <- as.matrix(dist(coords))
# predict movement between sites 3 and 4 using the radiation model
T_ij <- gravitywithdistance.flux(3, 4, d, pop, theta=c(1e-4,0.6,0.3,3,1,1e-4,0.6,0.3,3))
T_ij
}
\references{
Viboud et al. (2006) Synchrony, Waves, and Spatial Hierarchies in the Spread
of Influenza. \emph{Science} \url{http://dx.doi.org/10.1126/science.1125237}
}
\seealso{
\code{\link{movement.predict}}
}

