% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{intervening.opportunities}
\alias{intervening.opportunities}
\title{Intervening opportunities}
\usage{
intervening.opportunities(params = c(theta = 0.001, L = 1e-05))
}
\arguments{
\item{params}{A list of model parameters.  The following limits apply for the parameters: theta = [0, Inf]
and L = [0, Inf].}
}
\value{
A flux model object with the \code{\link{continuum.flux}} function and a set of starting parameters.
}
\description{
TODO
}
\note{
Limits \eqn{0} and \eqn{Inf} will be changed internally to the numerically safe approximations
\eqn{0 -> sqrt(.Machine$double.eps)} and \eqn{Inf -> sqrt(.Machine$double.xmax)}, respectively.
}
\seealso{
\code{\link{movement}}, \code{\link{continuum.flux}}, \code{\link{original.radiation}},
\code{\link{radiation.with.selection}}, \code{\link{uniform.selection}}, \code{\link{gravity}},
\code{\link{gravity.with.distance}}
}

