% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{is.location_dataframe}
\alias{is.location_dataframe}
\title{Check if given data.frame is 'location_dataframe' object}
\usage{
is.location_dataframe(x)
}
\arguments{
\item{x}{The object to be checked.}
}
\value{
True if the given object is a \code{location_dataframe} object; false otherwise
}
\description{
Helper function checking that the given object inherits from \code{location_dataframe} class.
}

