% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{is.locationdataframe}
\alias{is.locationdataframe}
\title{Check if given data.frame is 'locationdataframe' object}
\usage{
is.locationdataframe(x)
}
\arguments{
\item{x}{The object to be checked.}
}
\value{
True if the given object is a \code{locationdataframe} object; false otherwise
}
\description{
Helper function checking that the given object inherits from \code{locationdataframe} class.
}

