% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{is.movement_matrix}
\alias{is.movement_matrix}
\title{Check if given matrix if 'movement_matrix' object}
\usage{
is.movement_matrix(x)
}
\arguments{
\item{x}{The object to be checked.}
}
\value{
True if the given object is a \code{movement_matrix} object; false otherwise
}
\description{
Helper function checking that the given object inherits from \code{movement_matrix} class.
}

