% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{is.movementmatrix}
\alias{is.movementmatrix}
\title{Check if given matrix if 'movementmatrix' object}
\usage{
is.movementmatrix(x)
}
\value{
True if the given object is a \code{movementmatrix} object; false otherwise
}
\description{
Helper function checking that the given object inherits from \code{movementmatrix} class.
}

