% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{original.radiation}
\alias{original.radiation}
\title{Radiation model}
\usage{
original.radiation(params = c(theta = 0.9))
}
\arguments{
\item{params}{A list of model parameters. The limits for theta are [0,Infinity].}
}
\value{
A flux model object with the \code{\link{continuum.flux}} function and a set of starting parameters.
}
\description{
The (original) radiation model generally assumes the rational of job selection. It follows the general
rule that the number of employment opportunities in each district is proportional to its resident
population, assuming full employment (people in district = jobs in district). Moreover, the individuals
in each district choose the closest job to their home. Analytically the radiation model is represented by:
\deqn{T_{ij} = {\frac{PQ}{(P + R) (P + Q + R)}}}{T_ij = P * Q / (P + R) * (P + Q + R)}
where \eqn{P} is the population at the origin and \eqn{Q} at the destination, \eqn{R} denotes the total
population in a radius \eqn{\gamma} around population centres \eqn{P_i} and \eqn{Q_j}
}
\note{
For numerical reasons, the limits for theta will be internally changed to [eps, Infinity - eps].
TODO: will later on add a translist for the constraints of the parameters: theta-> positive values;
}
\references{
Simini, F., Gonzalez, M.C., Maritan, A. & Barabasi, A.-L. (2012). A universal model for mobility and
migration patterns. \emph{Nature}, 484, 96-100.
}
\seealso{
\code{\link{movement}}, \code{\link{continuum.flux}}
}

