% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{originalRadiation}
\alias{originalRadiation}
\title{Radiation model}
\usage{
originalRadiation(theta = 0.9)
}
\arguments{
\item{theta}{Model parameter \code{theta} with default value and the limits theta = [0, Inf].}
}
\value{
A flux model object with the \code{original radiation flux} function and a set of starting parameters.
}
\description{
The (original) radiation model generally assumes the rational of job selection. It follows the general
rule that the number of employment opportunities in each district is proportional to its resident
population, assuming full employment (people in district = jobs in district). Moreover, the individuals
in each district choose the closest job to their home. Analytically the radiation model is represented by:
\deqn{T_{ij} = {\frac{PQ}{(P + R) (P + Q + R)}}}{T_ij = P * Q / (P + R) * (P + Q + R)}
where \eqn{P} is the population at the origin and \eqn{Q} at the destination, \eqn{R} denotes the total
population in a radius \eqn{\gamma} around population centres \eqn{P_i} and \eqn{Q_j}.
}
\note{
Limits \eqn{0} and \eqn{Inf} will be changed internally to the numerically safe approximations
\eqn{0 -> sqrt(.Machine$double.eps)} and \eqn{Inf -> sqrt(.Machine$double.xmax)}, respectively.
}
\references{
Simini, F., Gonzalez, M.C., Maritan, A. & Barabasi, A.-L. (2012). A universal model for mobility and
migration patterns. \emph{Nature}, 484, 96-100.
}
\seealso{
\code{\link{movement}}, \code{\link{radiationWithSelection}}, \code{\link{uniformSelection}},
\code{\link{interveningOpportunities}}, \code{\link{gravity}}, \code{\link{gravityWithDistance}}
}

