% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{originalRadiationFlux}
\alias{originalRadiationFlux}
\title{Use the original radiation model of Simini et al. (2013) to predict movement between
two sites based on population and distance.}
\usage{
originalRadiationFlux(i, j, distance, population, theta = c(1),
  symmetric = FALSE, minpop = 1, maxrange = Inf)
}
\arguments{
\item{i}{Index for \code{population} and \code{distance} giving the first
site}

\item{j}{Index for \code{population} and \code{distance} giving the second
site}

\item{distance}{A distance matrix giving the euclidean distance between
pairs of sites}

\item{population}{A vector giving the population at all sites}

\item{theta}{A vector of the parameter which reflects the proportion of all
inhabitants in the region commuting}

\item{symmetric}{Whether to return a single value giving the total predicted
movements from i to j and j to i (if \code{TRUE}) or vector of length 2
giving movements from i to j (first element) and from j to i (second
element)}

\item{minpop}{The minimum population size to consider (by default 1,
consider all sites)}

\item{maxrange}{The maximum distance between sites to consider (by default
\code{Inf}, consider all sites)}
}
\value{
A vector (of length either 1 or 2) giving the predicted number of
people moving between the two sites.
}
\description{
Given indices \code{i} and \code{j}, a (dense) distance matrix
\code{distance} giving the euclidean distances between all pairs of sites, a
vector of population sizes \code{population} and a set of parameters, the
original radiation model as variant of the continuum model (Simini et al. 2013)
is used to predict movements between sites \code{i} and \code{j}. The mathematical definition
of the model and an explanation of how further continuum models are
related to each other can be found in Simini et al. (2013).
The parameter, which is required, is supplied as the first element of
the vector \code{theta}. This parameter describes the proportion of all
inhabitants in the region commuting. The default is that everyone commutes
and thus \code{theta[1]=1}.
The flux can be calculated either for both directions (by setting
\code{symmetric = FALSE}, returning movements for each direction) or for the
summed movement between the two (\code{symmetric = TRUE}).
The model can be speed up somewhat by setting \code{minpop} and
\code{maxrange}. If either of the two sites has a population lower than
\code{minpop} (minimum population size), or if the distance between the two
sites is greater than \code{maxrange} (the maximum range) it is assumed that
no travel occurs between these points.
Note that this function only works for individual site pairs. To calculate
movements across a whole landscape, use \code{\link{movement.predict}}.
}
\examples{
# generate random coordinates and populations
n <- 30
coords <- matrix(runif(n * 2), ncol = 2)
pop <- round(runif(n) * 1000)
# calculate the distance between pairs of sites
d <- as.matrix(dist(coords))
# predict movement between sites 3 and 4 using the original radiation model
T_ij <- originalRadiationFlux(3, 4, d, pop)
T_ij
}
\references{
Simini F, Maritan A, Neda Z (2013) Human mobility in a continuum approach.
\emph{PLoS ONE} 8(3): e60069.
\url{http://dx.doi.org/10.1371/journal.pone.0060069}
}
\seealso{
\code{\link{movement.predict}}, \code{\link{radiationWithSelectionFlux}},
\code{\link{uniformSelectionFlux}}, \code{\link{interveningOpportunitiesFlux}}
}

