% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{plot.movement_model}
\alias{plot.movement_model}
\title{Plot a movement model object}
\usage{
\method{plot}{movement_model}(x, ...)
}
\arguments{
\item{x}{a \code{movement_model}}

\item{\dots}{further arguments to be passed to or from other methods.}
}
\description{
plot 3 different figures ...
}

