% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict.movement_model}
\alias{predict.movement_model}
\title{Predict from a movement model object}
\usage{
\method{predict}{movement_model}(object, newdata, ...)
}
\arguments{
\item{object}{A configured prediction model of class \code{movement_model}}

\item{newdata}{An optional \code{location_dataframe} object or RasterLayer
containing population data}

\item{\dots}{Extra arguments to pass to the flux function}
}
\value{
A \code{movementpredictions} object which contains a list of the location
dataframe from the input and the matrix containing the predicted population movements.
}
\description{
\code{movement_model}:
Use a trained \code{movement_model} object to predict population movements
given either a RasterLayer containing a single population layer, or a
\code{location_dataframe}  object containing population and location data
with the columns \code{location} (character), \code{population} (numeric),
\code{x} (numeric) and \code{y} (numeric).
}

