% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict.movementmodel}
\alias{predict.movementmodel}
\title{Predictions from movementmodel objects}
\usage{
\method{predict}{movementmodel}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A configured prediction model of class \code{optimisedmodel}, ??}

\item{newdata}{An optional data.frame or RasterLayer containing population data}

\item{\dots}{Extra arguments to pass to the flux function}
}
\value{
A \code{movementmodel} containing a (dense) matrix giving predicted
movements between all sites. \code{optimisedmodel}: A list containing a location dataframe from the input, and a matrix
containing the predicted population movements.
}
\description{
Given a movement model, use the configured distances and
flux function to predict movement between all sites.
Any extra arguments of the flux functions can specified using the
\code{dots} argument.
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10,
                                 min_network_pop = 50000,
                                 predictionmodel= 'original radiation',
                                 symmetric = TRUE,
                                 modelparams = 0.1)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
sp::plot(raster::raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}

