% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict.optimisedmodel}
\alias{predict.optimisedmodel}
\title{Predict population movements from a population raster}
\usage{
\method{predict}{optimisedmodel}(model, raster)
}
\arguments{
\item{model}{An \code{optimisedmodel} object containing the trained model}

\item{raster}{A \code{RasterLayer} object containing a single population
attribute}
}
\value{
A list containing a location dataframe from the raster, and a matrix
containing the predicted population movements.
}
\description{
Use a trained \code{optimisedmodel} object to predict population movements
given a raster containing a single population layer.
}
\seealso{
\code{\link{movement}}, \code{\link{predict.movementmodel}}
}

