% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict}
\alias{predict}
\alias{predict.default}
\alias{predict.movementmodel}
\title{Create a movement prediction based on a configured movement model}
\usage{
predict(predictionModel, dataframe, ...)

\method{predict}{default}(predictionModel, dataframe, ...)

\method{predict}{movementmodel}(predictionModel, dataframe = NULL, ...)
}
\arguments{
\item{predictionModel}{A configured prediction model}

\item{dataframe}{An optional data frame containing population data}

\item{\dots}{Extra arguments to pass to the flux function}
}
\value{
A \code{movementmodel} containing a (dense) matrix giving predicted
movements between all sites.
}
\description{
Create a movement prediction based on a configured movement model
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default action for predict

\item \code{movementmodel}: Given a movement model, use the configured distances and
flux function to predict movement between all sites.
Any extra arguments of the flux functions can specified using the
\code{dots} argument.
}}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10,
                                 min_network_pop = 50000,
                                 predictionmodel= 'original radiation',
                                 symmetric = TRUE,
                                 modelparams = 0.1)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
sp::plot(raster::raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}
\seealso{
\code{\link{movementmodel}}, \code{\link{showprediction}}
}

