% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{print.movement_model}
\alias{print.movement_model}
\title{Print a movement model object}
\usage{
\method{print}{movement_model}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{an \code{movement_model} object}

\item{digits}{number of significant digits, see print.default.}

\item{\dots}{further arguments to be passed to or from other methods.
They are ignored in this function.}
}
\description{
Print etails of an optimised movement model
}

