% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{radiation.with.selection}
\alias{radiation.with.selection}
\title{Radiation with selection model}
\usage{
radiation.with.selection(params = c(theta = 0.1, lambda = 0.2))
}
\arguments{
\item{params}{A list of model parameters.  The following limits apply for the parameters: theta = [0, Inf]
and lambda = [0,1].}
}
\value{
A flux model object with the \code{\link{continuum.flux}} function and a set of starting parameters.
}
\description{
The aim of this index is to represent travel from districts between the affected countries to other
districts within the core countries. We assume that travel between districts is determined by factors
such as population and distance. The radiation model with selection is defined as:
\deqn{ T_{ij} = \frac{\frac{1 - \lambda^{P}}{P} - \frac{1 - \lambda^{Q}}{Q}}{\frac{1 - \lambda^{R}}{R}} }{%
T_ij = ( 1 - \lambda^P / P ) *  ( 1 - \lambda^Q / Q ) / ( 1 - \lambda^R / R )}
where \eqn{P} is the population at the origin and \eqn{Q} at the destination, \eqn{R} denotes the total
population in a radius \eqn{\gamma} around population centres \eqn{P_i} and \eqn{Q_j}.
The radiation model with selection was fitted using a set of known between district (n = 329) movements
from mobile phone users from France in 2007 (Tizzoni et al. 2014). The model was then used to build a
movement matrix between all districts of the core countries. District level population data were extracted
using WorldPop. District level administrative boundaries were downloaded from GADM.
}
\note{
Limits \eqn{0} and \eqn{Inf} will be changed internally to the numerically safe approximations
\eqn{0 -> sqrt(.Machine$double.eps)} and \eqn{Inf -> sqrt(.Machine$double.xmax)}, respectively.
}
\references{
Simini, F., Gonzalez, M.C., Maritan, A. & Barabasi, A.-L. (2012). A universal model for mobility and
migration patterns. \emph{Nature}, 484, 96-100.
Simini, F., Maritan, A. & Neda, Z. (2013). Human mobility in a continuum approach. \emph{PLoS One}, 8, e60069.
Tizzoni, M., Bajardi, P., Decuyper, A., Kon Kam King, G., Schneider, C.M., Blondel, V., et al. (2014).
On the Use of Human Mobility Proxies for Modeling Epidemics. \emph{PLoS Comput. Biol.}, 10, e1003716.
}
\seealso{
\code{\link{movement}}, \code{\link{continuum.flux}}, \code{\link{original.radiation}},
\code{\link{uniform.selection}}, \code{\link{intervening.opportunities}}, \code{\link{gravity}},
\code{\link{gravity.with.distance}}
}

