% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{show.prediction}
\alias{show.prediction}
\title{Display the movement predictions on a plot}
\usage{
show.prediction(network, predictedMovements)
}
\arguments{
\item{network}{A list containing a population vector, distance matrix and
sets of coordinates for each location}

\item{predictedMovements}{A data.frame containing predicted movements
between locations.}

\item{raster_layer}{A base layer to plot the predictions onto.}

\item{\dots}{Extra parameters to pass to plot}
}
\description{
Display the movement predictions on a plot
}

