% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{showcomparisonplot}
\alias{showcomparisonplot}
\title{Show a plot comparing and optimised model, and an observed dataset.}
\usage{
showcomparisonplot(optimisedmodel, observed)
}
\arguments{
\item{optimisedmodel}{An \code{OptimisedModel} object containing a trained
dataset.}

\item{observed}{An observed movement matrix}
}
\description{
Plots the observed movement matrix, predicted movement matrix and the
difference between the two.
}

