% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{showprediction}
\alias{showprediction}
\title{Display the movement predictions on a plot}
\usage{
showprediction(object, ...)
}
\arguments{
\item{object}{A configured prediction model}

\item{\dots}{Extra parameters to pass to plot}
}
\description{
Display the movement predictions on a plot
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10,
                                 min_network_pop = 50000,
                                 predictionmodel= 'original radiation',
                                 symmetric = TRUE,
                                 modelparams = 0.1)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
sp::plot(raster::raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}
\seealso{
\code{\link{movementmodel}}, \code{\link{predict.movementmodel}}
}

