% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R
\name{simplifytext}
\alias{simplifytext}
\title{Standardise a text string to upper case ASCII with no spaces}
\usage{
simplifytext(string)
}
\arguments{
\item{string}{The input string to simplify}
}
\value{
A standardised text string.
}
\description{
Standardise a text string to upper case ASCII with no spaces
}

