% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{summary.movement_model}
\alias{summary.movement_model}
\title{Summarize a movement model object}
\usage{
\method{summary}{movement_model}(object, ...)
}
\arguments{
\item{object}{a \code{movement_model} object}

\item{\dots}{additional arguments affecting the summary produced.}
}
\description{
Print a summary of a optimised movement model
}

