% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Function to convert a string phrase into a camelCase notation}
\usage{
toCamelCase(phrase, split, ...)
}
\arguments{
\item{phrase}{Character vector, each element of which is to be split. Other inputs, including a factor,
will give an error.}

\item{split}{Character vector (or object which can be coerced to such) containing regular expression(s)
(unless fixed = TRUE) to use for splitting. If empty matches occur, in particular if split has length 0,
phrase is split into single characters. If split has length greater than 1, it is re-cycled along phrase}

\item{\dots}{additional parameters which are passed to the \code{\link{strsplit}} function.}
}
\value{
A string in camelCase notation.
}
\description{
Convert a given string phrase into the camelCase notation using the \code{\link{strsplit}}
function.
}
\seealso{
\code{\link{strsplit}}
}

