% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{travelTime}
\alias{travelTime}
\title{Calculate travel times between coordinates}
\usage{
travelTime(friction, coords, coords2 = NULL, directions = 8, ...)
}
\arguments{
\item{friction}{a \code{RasterLayer} object with cell values giving the time
taken to traverse that cell.}

\item{coords}{a two-column matrix or dataframe, or a SpatialPoints object,
giving coordinates to compute travel time from. If \code{coords2 = NULL},
the a square matrix will be returned, giving travel time between these pairs
of coordinates}

\item{coords2}{an optional two-column matrix or dataframe, or a SpatialPoints
object, giving coordinates to compute travel time to.}

\item{directions}{directions in which cells are connected, can be either 4,
8, 16 or some other number. See \code{\link[gDistance]{adjacent}} for
details}

\item{\dots}{additional arguments to pass to
 \code{\link[gDistance]{transition}}.}
}
\value{
a matrix, with dimensions \code{nrow(coords), nrow(coords)}
 (If \code{coords2 = NULL}) or dimensions \code{nrow(coords), nrow(coords2)}
 otherwise
}
\description{
Using a friction raster (giving time taken to cross each cell)
 calculate the time taken to travel between pairs of coordinates.
}
\details{
This is a thin wrapper around functionality in the \code{gdistance}
 R package to facilitate the constuction of travel time matrices for use in
 movement models.
}
\examples{
# create a dummy friction matrix, assuming travel time is
# inversely proportional to population density
data(kenya)
kenya10 <- raster::aggregate(kenya, 10, sum)
friction <- 1 / kenya10

# example coordinates
coords <- matrix(c(38, 37, 36, 1, 0, 2), ncol = 2)

# calculate travel time between them
tt <- travelTime(friction, as.data.frame(coords))
tt
}

