% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{analysepredictionusingdpois}
\alias{analysepredictionusingdpois}
\title{Calculate the log likelihood of the prediction given the observed data.}
\usage{
analysepredictionusingdpois(prediction, observed)
}
\arguments{
\item{prediction}{A square matrix containing the predicted movements between location IDs}

\item{observed}{A square matrix containing the observed movements between location IDs}
}
\value{
The log likelihood
}
\description{
Processes an observed and predicted matrix, strips out the diagonals (which
should be zero) and calculates the log likelihood.
}
\seealso{
\code{\link{movementmodel}}, \code{\link{attemptoptimisation}}
}

