% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.data.frame.movement_matrix}
\alias{as.data.frame.movement_matrix}
\title{Convert a movement_matrix object into a data.frame}
\usage{
\method{as.data.frame}{movement_matrix}(x, ...)
}
\arguments{
\item{x}{a \code{movement_matrix} object.}

\item{\dots}{additional arguments to be passed to or from methods.}
}
\value{
A \code{data.frame} with columns \code{origin} (character), \code{destination} (character)
and \code{movement} (numeric)
}
\description{
Convert a \code{movement_matrix} object into a \code{data.frame} with
columns \code{origin} (character), \code{destination} (character) and \code{movement} (numeric).
The origins and the destinations are taken from the row and column names of the matrix,
respectively. The entries where origins are equal to destinations (that is, the matrix
diagonal) are removed during this conversion process.
}
\note{
If the given matrix does not contain any row or column names, the function will generate
a warning and the origin and destinations will be the row and column numbers of the relevant
matrix cell.
}

