% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.locationdataframe}
\alias{as.locationdataframe}
\alias{as.locationdataframe.SpatialPolygonsDataFrame}
\alias{as.locationdataframe.data.frame}
\title{Conversion to locationdataframe}
\usage{
as.locationdataframe(input, ...)

\method{as.locationdataframe}{data.frame}(input, ...)

\method{as.locationdataframe}{SpatialPolygonsDataFrame}(input, populationraster,
  ...)
}
\arguments{
\item{input}{object to convert to a \code{locationdataframe} object.
Either a data.frame with columns \code{origin} (character), \code{destination} (character), \code{movement} (numeric),
\code{pop_origin} (numberic), \code{pop_destination} (numberic), \code{lat_origin} (numberic), \code{long_origin} (numberic),
\code{lat_destination} (numberic) and \code{long_destination} (numberic), such as:
\tabular{ccccccccc}{
 origin \tab destination \tab movement \tab pop_origin \tab  pop_destination \tab  lat_origin \tab long_origin \tab lat_destination \tab long_destination \cr
 a   \tab     b    \tab    10   \tab     100     \tab      88          \tab 0.07826932  \tab 0.13612404  \tab     0.12114115   \tab    0.58984725 \cr
 a   \tab     c    \tab     8   \tab     100     \tab      100         \tab 0.07826932  \tab 0.13612404  \tab    0.07126503    \tab   0.19544754 \cr
 a   \tab     d    \tab    10   \tab     100     \tab      113         \tab 0.07826932  \tab 0.13612404  \tab     0.97817937   \tab    0.22771625 \cr
 a   \tab     e    \tab    11   \tab     100     \tab      107         \tab 0.07826932  \tab 0.13612404  \tab     0.87233335   \tab    0.06695538 \cr
 a   \tab     f    \tab    8    \tab     100     \tab      67          \tab 0.07826932  \tab  0.13612404 \tab    0.23157835    \tab   0.19573021 \cr
}
or a \code{SpatialPolygonsDataFrame} object}

\item{populationraster}{a \code{RasterLayer} object with each the value of
each cell giving the associated population density.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A data.frame containing location data of the format
#   location pop        lat        lon
# 1        a 100 0.07826932 0.13612404
# 2        b  88 0.12114115 0.58984725
# 3        c 100 0.07126503 0.19544754
# 4        d 113 0.97817937 0.22771625
# 5        e 107 0.87233335 0.06695538
}
\description{
Convert objects to \code{locationdataframe} objects
}

