% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.movement_matrix}
\alias{as.movement_matrix}
\alias{as.movement_matrix.data.frame}
\alias{as.movement_matrix.matrix}
\title{Conversion to movement_matrix}
\usage{
as.movement_matrix(object, ...)

\method{as.movement_matrix}{data.frame}(object, ...)

\method{as.movement_matrix}{matrix}(object, ...)
}
\arguments{
\item{object}{object to convert to a \code{movement_matrix} object.
Either a \code{data.frame} with columns \code{origin} (character), \code{destination} (character) and
\code{movement} (numeric) or a square \code{matrix} object}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A \code{movement_matrix} containing the observed movements.
}
\description{
Convert a \code{data.frame} or \code{matrix} object to a \code{movement_matrix}
object.


}

