% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.movementmatrix}
\alias{as.movementmatrix}
\title{Convert a data.frame into a movement matrix}
\usage{
as.movementmatrix(dataframe)
}
\arguments{
\item{dataframe}{A data.frame with the columns \code{origin} (character),
\code{destination} (character) and \code{movement} (numeric).}
}
\value{
A square matrix
}
\description{
Takes a dataframe listing movements between different locations and converts
it into a square matrix using the same location ids.
The dataframe does not need to include the a->a transitions as these are
automatically filled with zero if missing. This results in a zero diagonal
through the matrix.
}

