% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R
\name{as.movementmatrix}
\alias{as.movementmatrix}
\title{Convert a data.frame into a movement matrix}
\usage{
as.movementmatrix(dataframe)
}
\arguments{
\item{dataframe}{A data.frame of the format
#   origin destination movement
# 1      a           b       10
# 2      a           c        8
# 3      a           d       10
# 4      a           e       11
# 5      a           f        8
(truncated)}
}
\value{
A square matrix
}
\description{
Takes a dataframe listing movements between different locations and converts
it into a square matrix using the same location ids.
The dataframe does not need to include the a->a transitions as these are
automatically filled with zero if missing. This results in a zero diagonal
through the matrix.
}

