\name{createpopulationfromcsv}
\alias{createpopulationfromcsv}

\title{
  Create a data frame of populations at particular coordinates
}

\description{
Reads a correctly formatted csv file and creates a dataframe
}

\usage{
createpopulationfromcsv(filename)
}

\arguments{
  \item{filename}{File path of the csv file to process}
}

\value{
A dataframe containing csv data.
}

\seealso{
}
