% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{kenya}
\alias{kenya}
\title{Kenya 2010 population raster}
\format{A \code{RasterLayer} object in the WGS84 coordinate system at 1km
resolution.}
\source{
This is a product of the AfriPop project
\url{http://www.afripop.org} provided by Professor Andy Tatem.
}
\description{
An AfriPop raster of the modelled 2010 population in Kenya.
}
\examples{
\dontrun{
data(kenya)
sp::plot(kenya)
}
}
\references{
Linard C., Gilbert, M. Snow, R.W., Noor, A.M. & Tatem, A.J. (2010)
Population Distribution, Settlement Patterns and Accessibility across
Africa in 2010. PLoS ONE
\url{http://www.plosone.org/article/info:doi/10.1371/journal.pone.0031743}
}

