% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{movement-package}
\alias{movement-package}
\title{Modelling and Analysing Movement Data for Epidemiology}
\description{
Movement of humans and animals has a crucial role in the epidemiology of a
number of diseases. Movement data is increasingly available to
epidemiologists and its incorporation in models and maps of disease is
increasingly popular. This package is a collaborative effort to improve our
ability to analyses movement data and to build and apply epidemiological
movement models.
}
\author{
Nick Golding, Andrew Schofield, Moritz Kraemer and Alex T. Perkins
Maintainer: Nick Golding <nick.golding.research at gmail.com>
}
\seealso{
\code{\link{movement}} \code{\link{predict.flux}}, \code{\link{predict.optimisedmodel}}
\code{\link{showprediction}}, \code{\link{get.network}}, \code{\link{kenya}},
}

