\name{movement-package}
\alias{movement-package}
\alias{movement}
\docType{package}
\title{
Functions to analyse and model movement data for epidemiology.
}
\description{
Movement of humans and animals has a crucial role in the epidemiology of a number of diseases. Movement data is increasingly available to epidemiologists and its incorporation in models and maps of disease is increasingly popular. This package is a collaborative effort to improve our ability to analyses movement data and to build and apply epidemiological movement models.

}
\details{
\tabular{ll}{
Package: \tab movement\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-08\cr
License: \tab GPL (>=2)\cr
}

}


\seealso{
\code{\link{movement.model}}
}

\author{
Nick Golding + others (to be added)

Maintainer: Nick Golding <nick.golding@zoo.ox.ac.uk>
}