% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{movement}
\alias{movement}
\title{Create an optimised movement model}
\usage{
movement(formula, flux_model = gravity(), ...)
}
\arguments{
\item{formula}{A formula with one response (a \code{movement_matrix} object) and
one predictor (a \code{location_dataframe} object), e.g. movement_matrix ~ location_dataframe.
The \code{location_dataframe} object contains location data and the \code{movement_matrix}
object contains the observed population movements.}

\item{flux_model}{The name of the movement model to use. Currently supported
models are \code{\link{originalRadiation}}, \code{\link{radiationWithSelection}},
\code{\link{uniformSelection}}, \code{\link{interveningOpportunities}},
\code{\link{gravity}} and \code{\link{gravityWithDistance}}}

\item{\dots}{Extra parameters to be passed to the prediction code.}
}
\value{
An \code{optimisedmodel} object containing the training results,
and the optimisation results.
}
\description{
Uses the \code{\link{optim}} method to create an optimised model of
population movements.
}
\note{
The most likely format of the location data will be as a single
\code{data.frame} with the columns \code{location}, \code{population}, \code{lat} and
\code{long}. This can be extracted from a larger dataframe with
\code{\link{as.location_dataframe}}
The \code{movement_matrix} can be extracted from a list of movements
using \code{\link{as.movement_matrix}}. To check that the given objects are suitable,
the helper functions \code{\link{is.location_dataframe}} and \code{\link{is.movement_matrix}}
can be used.
}
\examples{
# get location data
data(kenya)
kenya10 <- raster::aggregate(kenya, 10, sum)
net <- getNetwork(kenya10, min = 50000)
locationData <- data.frame(location = net$locations, population = net$population, x = net$coordinate[,1], y = net$coordinate[,2])
class(locationData) <- c('data.frame', 'location_dataframe')
# simulate movements (note the values of movementmatrix must be integer)
predictedMovement  <- predict(originalRadiation(theta = 0.1), locationData, symmetric = TRUE)
movementMatrix <- predictedMovement$movement_matrix
# fit a new model to these data
s <- movement(movementMatrix ~ locationData, radiationWithSelection(theta = 0.5))
s
}
\seealso{
\code{\link{as.location_dataframe}}, \code{\link{is.location_dataframe}},
\code{\link{as.movement_matrix}}, \code{\link{is.movement_matrix}}, \code{\link{originalRadiation}},
\code{\link{radiationWithSelection}}, \code{\link{uniformSelection}},
\code{\link{interveningOpportunities}}, \code{\link{gravity}} and \code{\link{gravityWithDistance}}
}

