% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{movement}
\alias{movement}
\title{Create an optimised movement model}
\usage{
movement(formula, flux_model = gravity(), ...)
}
\arguments{
\item{formula}{A formula with one response (a \code{movementmatrix} object) and
one predictor (a \code{locationdataframe} object), e.g. movementmatrix ~ locationdataframe.
The \code{locationdataframe} object contains location data and the \code{movementmatrix}
object contains the observed population movements.}

\item{flux_model}{The name of the movement model to use. Currently supported
models are \code{\link{original.radiation}}, \code{\link{radiation.with.selection}},
\code{\link{uniform.selection}}, \code{\link{intervening.opportunities}},
\code{\link{gravity}} and \code{\link{gravity.with.distance}}}

\item{\dots}{Extra parameters to be passed to the prediction code.}
}
\value{
An \code{optimisedmodel} object containing the training results,
and the optimisation results.
}
\description{
Uses the \code{\link{optim}} method to create an optimised model of
population movements.
}
\note{
The most likely format of the location data will be as a single
\code{data.frame} with the columns \code{location}, \code{population}, \code{lat} and
\code{long}. This can be extracted from a larger dataframe with
\code{\link{as.locationdataframe}}
The \code{movement_matrix} can be extracted from a list of movements
using \code{\link{as.movementmatrix}}. To check that the given objects are suitable,
the helper functions \code{\link{is.locationdataframe}} and \code{\link{is.movementmatrix}}
can be used.
}
\examples{
# TODO: add new exmple using the movement() function
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
#predictionModel <- movementmodel(dataset=kenya10,
 #                                min_network_pop = 50000,
 #                                flux_model = original.radiation(),
 #                                symmetric = TRUE)
# predict the population movement from the model
#predictedMovements = predict(predictionModel)
# visualise the distance matrix
#sp::plot(raster::raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
#showprediction(predictedMovements)
}
\seealso{
\code{\link{as.locationdataframe}}, \code{\link{is.locationdataframe}},
\code{\link{as.movementmatrix}}, \code{\link{is.movementmatrix}}
}

