% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R, R/pkgname.R
\docType{package}
\name{movement}
\alias{movement}
\alias{movement-package}
\title{Create an optimised movement model}
\usage{
movement(locations, coords, population, movement_matrix, model)
}
\arguments{
\item{locations}{A vector containing populations}

\item{coords}{A data frame containing coordinates of the \code{locations}}

\item{population}{A vector containing populations of the \code{locations}}

\item{movement_matrix}{A square matrix containing the observed population
movements between \code{locations}}

\item{model}{The name of the movement model to use. Currently supported
models are \code{original radiation}, \code{radiation with selection},
\code{uniform selection}, \code{intervening opportunities},
\code{gravity}}
}
\value{
An \code{optimisedmodel} object containing the training results,
and the optimisation results. This can then be used by
\code{\link{predict}} to generate predictions on new data.
}
\description{
Uses the \code{\link{optim}} method to create an optimised model of
population movements.

Movement of humans and animals has a crucial role in the epidemiology of a
number of diseases. Movement data is increasingly available to
epidemiologists and its incorporation in models and maps of disease is
increasingly popular. This package is a collaborative effort to improve our
ability to analyses movement data and to build and apply epidemiological
movement models.
}
\details{
\tabular{ll}{
Package: \tab movement\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-08\cr
License: \tab GPL (>=2)\cr
}
}
\note{
The most likely format of the location data will be as a single
\code{data.frame} of this form:
#   location pop        lat        lon
# 1        a 100 0.07826932 0.13612404
# 2        b  88 0.12114115 0.58984725
# 3        c 100 0.07126503 0.19544754
# 4        d 113 0.97817937 0.22771625
# 5        e 107 0.87233335 0.06695538
This can be extracted from a larger dataframe with
\code{\link{as.locationdataframe}}
The \code{movement_matrix} can be extracted from a list of movements
using \code{\link{as.movementmatrix}}
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10, min_network_pop = 50000, predictionmodel= 'original radiation', symmetric = TRUE, modelparams = 0.1)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
plot(raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}
\author{
Nick Golding + others (to be added)
Maintainer: Nick Golding <nick.golding at zoo.ox.ac.uk>
}
\seealso{
\code{\link{predict}}, \code{\link{as.locationdataframe}},
\code{\link{as.movementmatrix}}

\code{\link{movementmodel}}, \code{\link{predict}},
\code{\link{showprediction}}, \code{\link{movement.predict}},
\code{\link{get.network}}, \code{\link{kenya}}, \code{\link{gravity.flux}},
\code{\link{continuum.flux}}, \code{\link{attemptoptimisation}}
}

