% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{movementmodel}
\alias{movementmodel}
\title{Create a movement model to predict movements across a landscape.}
\usage{
movementmodel(dataset, min_network_pop = 50000,
  flux.model = original.radiation(), symmetric = TRUE)
}
\arguments{
\item{dataset}{A raster dataset of population data}

\item{min_network_pop}{The minimum population of a site in order for it to be
processed}

\item{symmetric}{Whether to calculate symmetric or asymmetric (summed across
both directions) movement}

\item{predictionmodel}{The name of a prediction model used to calculated
predicted flux between locations. Currently supported prediction models are
\code{gravity}, \code{original radiation}, \code{intervening opportunities},
\code{radiation with selection} and \code{uniform selection}}

\item{modelparams}{Model parameter values to pass to the prediction model
(such as exponents, scale factors etc.)}
}
\value{
A movement model object which can be used to run flux predictions.
}
\description{
This sets up a movement model object which can then be used to predict
population movements. It requires a \code{raster} dataset and a number of
optional parameters to set up the basic configuration of the prediction
model.
The model can be calculated either for both directions (by setting
\code{symmetric = FALSE}, resulting in an asymmetric movement matrix) or for
the summed movement between the two (\code{symmetric = TRUE}, giving a
symmetric matrix)).
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10,
                                 min_network_pop = 50000,
                                 predictionmodel= 'original radiation',
                                 symmetric = TRUE,
                                 modelparams = 0.1)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
sp::plot(raster::raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}
\seealso{
\code{\link{predict.movementmodel}}, \code{\link{showprediction}}
}

