% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict.flux}
\alias{predict.flux}
\title{Predict from theoretical flux object}
\usage{
\method{predict}{flux}(object, locationdataframe, min_network_pop = 50000,
  symmetric = FALSE, ...)
}
\arguments{
\item{object}{A theoretical model of type \code{flux} object}

\item{locationdataframe}{A data.frame or RasterLayer containing population data}

\item{min_network_pop}{Optional parameter for the minimum population of a site
in order for it to be processed}

\item{symmetric}{Optional parameter to define whether to calculate symmetric or
asymmetric (summed across both directions) movement}

\item{\dots}{further arguments to be passed to or from other methods.
They are ignored in this function.}
}
\value{
A list containing a location dataframe from the input with columns
\code{location}, \code{population} and \code{coordinates} and a matrix
containing the predicted population movements.
}
\description{
Use a \code{flux} object to predict population movements
given either a RasterLayer containing a single population layer, or a
data.frame containing population and location data with the columns
\code{origin} (character), \code{pop_origin} (numeric), \code{long_origin} (numeric)
and \code{lat_origin} (numeric).

The model can be calculated either for both directions (by setting the optional parameter
\code{symmetric = FALSE}, resulting in an asymmetric movement matrix) or for
the summed movement between the two (\code{symmetric = TRUE}, giving a
symmetric matrix)).
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# generate a flux object
flux <- radiation.with.selection()
# run the prediction for the theoretical model
predictedMovement  <- predict(flux, kenya10)
}

