% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict.flux}
\alias{predict.flux}
\title{Predict from theoretical flux object}
\usage{
\method{predict}{flux}(object, location_dataframe, min_network_pop = 50000,
  symmetric = FALSE, go_parallel = FALSE, number_of_cores = NULL, ...)
}
\arguments{
\item{object}{A theoretical model of type \code{flux} object}

\item{location_dataframe}{A \code{location_dataframe} object or RasterLayer containing population data}

\item{min_network_pop}{Optional parameter for the minimum population of a site
in order for it to be processed}

\item{symmetric}{Optional parameter to define whether to calculate symmetric or
asymmetric (summed across both directions) movement}

\item{go_parallel}{Flag to enable parallel calculations (if set to TRUE).
Note that parallel programming will only improve the performance with larger datasets; for smaller
datasets the performance will get worse due to the overhead of scheduling the tasks.}

\item{number_of_cores}{Optional parameter to specify the number of cores used for parallel calculations.
If no value is specified, the program will automatically detect the number of cores available on the machine
when parallel programming is enabled.}

\item{\dots}{additional arguments affecting the predictions produced.}
}
\value{
A list containing a location dataframe from the input with columns
\code{location}, \code{population} and \code{coordinates} and a matrix
containing the predicted population movements.
}
\description{
Use a \code{flux} object to predict population movements
given either a RasterLayer containing a single population layer, or a
\code{location_dataframe}  object containing population and location data with the columns
\code{location} (character), \code{population} (numeric), \code{x} (numeric)
and \code{y} (numeric).

The model can be calculated either for both directions (by setting the optional parameter
\code{symmetric = FALSE}, resulting in an asymmetric movement matrix) or for
the summed movement between the two (\code{symmetric = TRUE}, giving a
symmetric matrix)).
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# generate a flux object
flux <- radiationWithSelection()
# run the prediction for the theoretical model
predicted_movement  <- predict(flux, kenya10)
}

