% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{predict.optimisedmodel}
\alias{predict.optimisedmodel}
\title{Predict from an optimisedmodel object}
\usage{
\method{predict}{optimisedmodel}(object, newdata, ...)
}
\arguments{
\item{object}{A configured prediction model of class \code{optimisedmodel}, ??}

\item{newdata}{An optional data.frame or RasterLayer containing population data}

\item{\dots}{Extra arguments to pass to the flux function}
}
\value{
A \code{movementmodel} containing a (dense) matrix giving predicted
movements between all sites. \code{optimisedmodel}: A list containing a location dataframe from the input, and a matrix
containing the predicted population movements.
}
\description{
\code{optimisedmodel}:
Use a trained \code{optimisedmodel} object to predict population movements
given either a RasterLayer containing a single population layer, or a
data.frame containing population and location data formatted as:
#   location pop        lat        lon
# 1        a 100 0.07826932 0.13612404
# 2        b  88 0.12114115 0.58984725
# 3        c 100 0.07126503 0.19544754
# 4        d 113 0.97817937 0.22771625
# 5        e 107 0.87233335 0.06695538
}

