% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{print.flux}
\alias{print.flux}
\title{Print details of a flux object}
\usage{
\method{print}{flux}(x, ...)
}
\arguments{
\item{x}{a \code{flux} object}

\item{\dots}{further arguments to be passed to or from other methods.
They are ignored in this function.}
}
\description{
Print details of a given flux object
}
\examples{
flux <- gravity(theta = 0.1, alpha = 0.5, beta = 0.1, gamma = 0.1)
print(flux)
}

