% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R
\name{rasterizeShapeFile}
\alias{rasterizeShapeFile}
\title{Convert a shapefile to a \code{RasterLayer}}
\usage{
rasterizeShapeFile(filename, keeplist)
}
\arguments{
\item{filename}{The path to the shapefile to convert}

\item{keeplist}{A list of layers in the shapefile to keep}
}
\value{
A \code{RasterLayer} object
}
\description{
Converts a given shapefile and converts it to a \code{RasterLayer} and
discards any unnecessary layers.
}

