% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{showprediction}
\alias{showprediction}
\alias{showprediction.default}
\alias{showprediction.movement_predictions}
\title{Plot movement predictions}
\usage{
showprediction(object, ...)

\method{showprediction}{default}(object, ...)

\method{showprediction}{movement_predictions}(object, ...)
}
\arguments{
\item{object}{A \code{movement_predictions} object}

\item{\dots}{Extra parameters to pass to plot}
}
\description{
Given a predicted movement model (that is a \code{movement_predictions} object),
plot the underlying raster, the configured location points and the predicted movements
between locations.
}

