% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{showprediction}
\alias{showprediction}
\alias{showprediction.default}
\alias{showprediction.movementmodel}
\title{Plot movement predictions}
\usage{
showprediction(object, ...)

\method{showprediction}{default}(object, ...)

\method{showprediction}{movementmodel}(object, ...)
}
\arguments{
\item{object}{A configured prediction model}

\item{\dots}{Extra parameters to pass to plot}
}
\description{
Given a movement model, plot the underlying
raster, the configured location points and the predicted movements
between locations.
}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- raster::aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10,
                                 min_network_pop = 50000,
                                 flux_model = original.radiation(),
                                 symmetric = TRUE)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
sp::plot(raster::raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}
\seealso{
\code{\link{movementmodel}}, \code{\link{predict.movementmodel}}
}

