% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{showprediction}
\alias{showprediction}
\alias{showprediction.default}
\title{Plot movement predictions}
\usage{
showprediction(object, ...)

\method{showprediction}{default}(object, ...)
}
\arguments{
\item{object}{A configured prediction model}

\item{\dots}{Extra parameters to pass to plot}
}
\description{
Given a movement model, plot the underlying
raster, the configured location points and the predicted movements
between locations.
}
\seealso{
\code{\link{movementmodel}}, \code{\link{predict.movementmodel}}
}

