% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/movement.R
\name{showprediction}
\alias{showprediction}
\alias{showprediction.default}
\alias{showprediction.movementmodel}
\title{Display the movement predictions on a plot}
\usage{
showprediction(predictionModel, ...)

\method{showprediction}{default}(predictionModel, ...)

\method{showprediction}{movementmodel}(predictionModel, ...)
}
\arguments{
\item{predictionModel}{A configured prediction model}

\item{\dots}{Extra parameters to pass to plot}
}
\description{
Display the movement predictions on a plot
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default action for showprediction

\item \code{movementmodel}: Given a movement model, plot the underlying
raster, the configured location points and the predicted movements
between locations.
}}
\examples{
# load kenya raster
data(kenya)
# aggregate to 10km to speed things up
kenya10 <- aggregate(kenya, 10, sum)
# create the prediction model for the aggregate dataset using the fixed parameter radiation model
predictionModel <- movementmodel(dataset=kenya10, min_network_pop = 50000, predictionmodel= 'original radiation', symmetric = TRUE, modelparams = 0.1)
# predict the population movement from the model
predictedMovements = predict(predictionModel)
# visualise the distance matrix
plot(raster(predictedMovements$net$distance_matrix))
# visualise the predicted movements overlaid onto the original raster
showprediction(predictedMovements)
}
\seealso{
\code{\link{movementmodel}}, \code{\link{predict}}
}

