% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{summary.movement_model}
\alias{summary.movement_model}
\title{Summarize a movement model object}
\usage{
\method{summary}{movement_model}(object, ...)
}
\arguments{
\item{object}{a \code{movement_model} object}

\item{\dots}{additional arguments affecting the summary produced.}
}
\value{
Returns a \code{summary.movement_model} object containing a list of the following parameters
\item{model}{The \code{flux} model objects used}
\item{deviance_resid}{the deviance residuals}
\item{coefficients}{The parameters estimates on the true scale used in the flux model equations}
\item{trans_coeff}{The parameters estimates on the continuous scale, after parameter transformation,
 used for the optimisation process}
\item{trans_coeff_std_errors}{The standard error of the optimised parameters}
\item{null_deviance}{The null deviance}
\item{df_null}{the degree of freedom on the null deviance}
\item{resid_deviance}{the residual deviance}
\item{df_residual}{the degree of freedom on the residual deviance}
\item{aic}{the aic}
}
\description{
Creates a summary of an optimised \code{movement_model} object.
}

