% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{summary.optimisedmodel}
\alias{summary.optimisedmodel}
\title{Summarize an optimised model

Print a summary of an optimised model}
\usage{
\method{summary}{optimisedmodel}(object, ...)
}
\arguments{
\item{object}{an \code{optimisedmodel} object}

\item{\dots}{additional arguments affecting the summary produced.}
}
\description{
Summarize an optimised model

Print a summary of an optimised model
}

