% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movement.R
\name{uniform.selection}
\alias{uniform.selection}
\title{Uniform selection model}
\usage{
uniform.selection(theta = 0.9)
}
\arguments{
\item{theta}{Model parameter with default value and the limits theta = [0, Inf].}
}
\value{
A flux model object with the \code{\link{uniformSelectionFlux}} function and a set of starting parameters.
}
\description{
The uniform selection model assumes that a job is selected uniformly at random proportionally to the
population in each district following:
\deqn{T_{ij} = \frac{P}{Q-R}}{T_ij = P / Q - R}
where \eqn{P} is the population at the origin and \eqn{Q} at the destination, \eqn{R} denotes the total
population in a radius \eqn{\gamma} around population centres \eqn{P_i} and \eqn{Q_j}.
}
\note{
Limits \eqn{0} and \eqn{Inf} will be changed internally to the numerically safe approximations
\eqn{0 -> sqrt(.Machine$double.eps)} and \eqn{Inf -> sqrt(.Machine$double.xmax)}, respectively.
}
\references{
Simini, F., Maritan, A. & Neda, Z. (2013). Human mobility in a continuum approach. \emph{PLoS One}, 8, e60069.
}
\seealso{
\code{\link{movement}}, \code{\link{uniformSelectionFlux}}, \code{\link{original.radiation}},
\code{\link{radiation.with.selection}}, \code{\link{intervening.opportunities}}, \code{\link{gravity}},
\code{\link{gravity.with.distance}}
}

